dps: Make changes for the DPS5005 with v1.4 firmware
Ultimately these changes should be abstracted out for various models
and firmwares. I don't know which devices and firmwares have which
precisions so for now I am making the changes specific to my device.

The specific changes are increasing the modifier for the Amps as the
DPS5005 supports 3 decimal precision, and the modifier on the firmware
which is major.minor. Then for human readable format I have changed the
formating to properly display this.

Signed-off-by: Jon Nettleton <jon@solid-run.com>